separated components from containers, linted the code
