+++ src/components/App.js
@@ -5,51 +5,47 @@ import { connect } from 'react-redux';
 import { withRouter, Route } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import Welcome from './Welcome';
-import MealSearch from './MealSearch';
-import Meal from './Meal';
+import MealSearch from '../containers/MealSearch';
+import Meal from '../containers/Meal';
 import CategorySearch from './CategorySearch';
 import * as actions from '../actions/actions';
 
-// eslint-disable-next-line react/prefer-stateless-function
-class App extends React.Component {
-  render() {
-    const {
-      state, changeCategory, addMeals, addMeal, searchMeal,
-    } = this.props;
-    const { categoryReducer } = state;
-    return (
-      <div className="app-container">
-        <Route
-          exact
-          path="/"
-          render={() => (
-            <div>
-              <Welcome />
-              <MealSearch addMeal={addMeal} state={state} searchMeal={searchMeal} />
-              <CategorySearch
-                state={state}
-                changeCategory={changeCategory}
-                categoryReducer={categoryReducer}
-                addMeals={addMeals}
-              />
-            </div>
-          )}
-        />
-        <Route
-          path="/meal/:id"
-          render={params => (
-            <div>
-              <Meal addMeal={addMeal} match={params.match} state={state} />
-            </div>
-          )}
-        />
-      </div>
-    );
-  }
+function App(props) {
+  const {
+    state, changeCategory, addMeals, addMeal, searchMeal,
+  } = props;
+  const { categoryReducer } = state;
+  return (
+    <div className="app-container">
+      <Route
+        exact
+        path="/"
+        render={() => (
+          <div>
+            <Welcome />
+            <MealSearch addMeal={addMeal} state={state} searchMeal={searchMeal} />
+            <CategorySearch
+              state={state}
+              changeCategory={changeCategory}
+              categoryReducer={categoryReducer}
+              addMeals={addMeals}
+            />
+          </div>
+        )}
+      />
+      <Route
+        path="/meal/:id"
+        render={params => (
+          <div>
+            <Meal addMeal={addMeal} match={params.match} state={state} />
+          </div>
+        )}
+      />
+    </div>
+  );
 }
 
 App.propTypes = {
-  // eslint-disable-next-line react/forbid-prop-types
   state: PropTypes.shape({
     categoryReducer: PropTypes.string.isRequired,
     multiMealReducer: PropTypes.arrayOf(PropTypes.object),
