+++ src/components/CategorySearch.js
@@ -1,49 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import RenderCategs from './RenderCategs';
+import RenderCategs from '../containers/RenderCategs';
 import '../stylesheets/CategorySearch.css';
 
 const categoriesA = ['Beef', 'Chicken', 'Dessert', 'Lamb', 'Miscellaneous', 'Pasta', 'Pork', 'Seafood', 'Side', 'Starter', 'Vegan', 'Vegetarian', 'Breakfast', 'Goat'];
 let meals;
 
-class CategorySearch extends React.Component {
-  constructor() {
-    super();
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleChange(e) {
-    const { changeCategory } = this.props;
+function CategorySearch(props) {
+  const handleChange = e => {
+    const { changeCategory } = props;
     changeCategory(e.target.value);
-  }
+  };
 
-  render() {
-    const { categoryReducer, state, addMeals } = this.props;
-    if (categoryReducer === '') {
-      return (
-        <div className="categ-search">
-          <h2 className="categ-heading">Category Search</h2>
-          <form>
-            <label htmlFor="category">
-              Choose a category to search by
-              {' '}
-              <br />
-              <select onChange={this.handleChange} className="categs-list" name="category">
-                {categoriesA.map((categ, index) => (
-                  <option
-                    value={`${categ}`}
-                    // eslint-disable-next-line react/no-array-index-key
-                    key={`${index}`}
-                  >
-                    {`${categ}`}
-                  </option>
-                ))}
-              </select>
-            </label>
-          </form>
-        </div>
-      );
-    }
+  const { categoryReducer, state, addMeals } = props;
+  if (categoryReducer === '') {
     return (
       <div className="categ-search">
         <h2 className="categ-heading">Category Search</h2>
@@ -52,13 +22,11 @@ class CategorySearch extends React.Component {
             Choose a category to search by
             {' '}
             <br />
-            <select onChange={this.handleChange} className="categs-list" name="category">
+            <select onChange={handleChange} className="categs-list" name="category">
               {categoriesA.map((categ, index) => (
-
                 <option
                   value={`${categ}`}
-                  // eslint-disable-next-line react/no-array-index-key
-                  key={`${index}`}
+                  key={`${index + 100}`}
                 >
                   {`${categ}`}
                 </option>
@@ -66,24 +34,46 @@ class CategorySearch extends React.Component {
             </select>
           </label>
         </form>
-        <div>
-          <RenderCategs
-            state={state}
-            addMeals={addMeals}
-            categoryReducer={categoryReducer}
-            meals={meals}
-          />
-        </div>
       </div>
     );
   }
+  return (
+    <div className="categ-search">
+      <h2 className="categ-heading">Category Search</h2>
+      <form>
+        <label htmlFor="category">
+          Choose a category to search by
+          {' '}
+          <br />
+          <select onChange={handleChange} className="categs-list" name="category">
+            {categoriesA.map((categ, index) => (
+
+              <option
+                value={`${categ}`}
+                key={`${index + 100}`}
+              >
+                {`${categ}`}
+              </option>
+            ))}
+          </select>
+        </label>
+      </form>
+      <div>
+        <RenderCategs
+          state={state}
+          addMeals={addMeals}
+          categoryReducer={categoryReducer}
+          meals={meals}
+        />
+      </div>
+    </div>
+  );
 }
 
 CategorySearch.propTypes = {
-  // eslint-disable-next-line react/forbid-prop-types
   state: PropTypes.shape({
     categoryReducer: PropTypes.string.isRequired,
-    multiMealReducer: PropTypes.arrayOf(PropTypes.object),
+    multiMealReducer: PropTypes.arrayOf(PropTypes.instanceOf(Object)),
     oneMealReducer: PropTypes.instanceOf(Object),
     searchMealReducer: PropTypes.string,
   }).isRequired,
